WIP: add custom socket to all http requests
