+++ src/lib/peer-store/PeerInfo.ts
@@ -23,7 +23,7 @@ class Address {
     return `${this.ip}:${this.port}`
   }
 
-  fromAddressInfo(addressInfo: AddressInfo) {
+  static fromAddressInfo(addressInfo: AddressInfo) {
     const ip = (addressInfo.address == '::') ? '127.0.0.1' : addressInfo.address
     return new Address(ip, addressInfo.port.toString())
   }
